refactor : 여러 알고리즘이 가능하도록 전략패턴 적용
