+++ src/main/java/wooteco/subway/admin/service/PathService.java
@@ -2,25 +2,18 @@ package wooteco.subway.admin.service;
 
 import static java.util.stream.Collectors.*;
 
-import java.util.Collection;
 import java.util.List;
 import java.util.Map;
-import java.util.NoSuchElementException;
-import java.util.Objects;
 
 import org.jgrapht.GraphPath;
-import org.jgrapht.alg.shortestpath.DijkstraShortestPath;
-import org.jgrapht.graph.WeightedMultigraph;
 import org.springframework.stereotype.Service;
 
 import wooteco.subway.admin.domain.Edge;
 import wooteco.subway.admin.domain.Line;
-import wooteco.subway.admin.domain.LineStation;
-import wooteco.subway.admin.domain.PathType;
 import wooteco.subway.admin.domain.Station;
+import wooteco.subway.admin.domain.strategy.AlgorithmStrategy;
 import wooteco.subway.admin.dto.PathRequest;
 import wooteco.subway.admin.dto.PathResponse;
-import wooteco.subway.admin.exception.NotConnectEdgeException;
 import wooteco.subway.admin.repository.LineRepository;
 import wooteco.subway.admin.repository.StationRepository;
 
@@ -34,62 +27,20 @@ public class PathService {
         this.stationRepository = stationRepository;
     }
 
-    public PathResponse findShortestPathByDistance(PathRequest pathRequest) {
+    public PathResponse findShortestPath(PathRequest pathRequest, AlgorithmStrategy algorithmStrategy) {
         final Map<Long, Station> stations = stationRepository.findAll()
             .stream()
             .collect(toMap(Station::getId, station -> station));
         final List<Line> lines = lineRepository.findAll();
 
-        final Station startStation = getStationWithValidation(stations, pathRequest.getSource());
-        final Station endStation = getStationWithValidation(stations, pathRequest.getTarget());
-
-        DijkstraShortestPath<Station, Edge> dijkstraShortestPath =
-            new DijkstraShortestPath<>(makeGraph(stations, lines, pathRequest.getType()));
-
-        final GraphPath<Station, Edge> path = dijkstraShortestPath.getPath(startStation, endStation);
-
-        if (Objects.isNull(path)) {
-            throw new NotConnectEdgeException(startStation, endStation);
-        }
+        final GraphPath<Station, Edge> path = algorithmStrategy.getPath(stations, lines, pathRequest);
 
         List<Station> shortestPath = path.getVertexList();
-        ;
         List<Edge> edgeList = path.getEdgeList();
 
         final int distance = edgeList.stream().mapToInt(Edge::getDistance).sum();
         final int duration = edgeList.stream().mapToInt(Edge::getDuration).sum();
 
         return new PathResponse(shortestPath, distance, duration);
-    }
-
-    private WeightedMultigraph<Station, Edge> makeGraph(Map<Long, Station> stations, List<Line> lines,
-        PathType pathType) {
-        WeightedMultigraph<Station, Edge> graph
-            = new WeightedMultigraph<>(Edge.class);
-
-        stations.values()
-            .forEach(graph::addVertex);
-
-        lines.stream()
-            .map(Line::getStations)
-            .flatMap(Collection::stream)
-            .filter(LineStation::isEdge)
-            .forEach(lineStation -> {
-                Station preStation = getStationWithValidation(stations, lineStation.getPreStationId());
-                Station currentStation = getStationWithValidation(stations, lineStation.getStationId());
-                Edge edge = lineStation.toEdge();
-
-                graph.addEdge(preStation, currentStation, edge);
-                graph.setEdgeWeight(edge, pathType.getValue(lineStation));
-            });
-        return graph;
-    }
-
-    private Station getStationWithValidation(Map<Long, Station> stations, Long stationId) {
-        if (!stations.containsKey(stationId)) {
-            throw new NoSuchElementException("해당 역은 등록되어있지 않은 역입니다.");
-        }
-
-        return stations.get(stationId);
     }
 }
